# contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Condux Control Plane API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8099/"
    }
  ],
  "paths": {
...
    "/api/orgs/{orgId}/invites": {
      "post": {
        "tags": [
          "Invites"
        ],
        "operationId": "createInvite",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Invites"
        ],
        "operationId": "listInvites",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitePendingResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orgs/{orgId}/invites/{inviteId}": {
      "delete": {
        "tags": [
          "Invites"
        ],
        "operationId": "revokeInvite",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "inviteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/invites/accept": {
      "post": {
        "tags": [
          "Invites"
        ],
        "operationId": "acceptInvite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgMembershipResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInviteRequest": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "CreateInviteRequest": {
        "required": [
          "email",
          "role"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "InvitePendingResponse": {
        "required": [
          "id",
          "email",
          "role",
          "createdAt",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InviteResponse": {
        "required": [
          "id",
          "email",
          "role",
          "expiresAt",
          "token"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "Org": {
        "required": [
          "id",
          "slug",
          "name",
          "tier",
          "createdAt",
          "aiFixMode",
          "aiFixCostCapUsd"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tier": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "aiFixMode": {
            "type": "integer",
            "format": "int32"
          },
          "aiFixCostCapUsd": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stripeCustomerId": {
            "type": "string",
            "nullable": true
          },
          "stripeSubscriptionId": {
            "type": "string",
            "nullable": true
          },
          "weeklySummaryEnabled": {
            "type": "boolean",
            "default": true
          },
          "weeklySummaryDow": {
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          "weeklySummaryHour": {
            "type": "integer",
            "format": "int32",
            "default": 9
          },
          "weeklySummaryTz": {
            "type": "string",
            "default": "UTC"
          }
        }
      },
...
      "OrgMembershipResponse": {
        "required": [
          "org",
          "role"
        ],
        "type": "object",
        "properties": {
          "org": {
            "$ref": "#/components/schemas/Org"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}